delete from cart modified

+++ src/Pages/Cart.jsx
@@ -3,6 +3,7 @@
 import { useState } from "react";
 import toast from "react-hot-toast";
 import { useLoaderData } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Cart = () => {
 
@@ -20,27 +21,46 @@ const Cart = () => {
 
     const handleDelete = (id) => {
         // make sure
-        fetch(`http://localhost:5000/cart/${id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
         })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount > 0){
-                console.log('deleted successfully');
-                toast.success('Product deleted from cart successfully!');
+            .then(result => {
+                if (result.isConfirmed) {
+                    fetch(`http://localhost:5000/cart/${id}`, {
+                        method: 'DELETE'
+                    })
+                        .then(res => res.json())
+                        .then(data => {
+                            if (data.deletedCount > 0) {
+                                const remainingItems = items.filter(item => item._id !== id);
+                                setItems(remainingItems);
 
-                const remainingItems = items.filter(item => item._id !==id);
-                setItems(remainingItems);
-            }
-        })
+                                
+                                console.log('deleted successfully');
+                                Swal.fire(
+                                    'Deleted!',
+                                    'Your Coffee has been deleted.',
+                                    'success'
+                                )
 
+                            }
+                        })
+                }
+            })
     }
+
 
     // console.log(items);
     return (
         <div className="min-h-screen">
             <h2 className='text-4xl font-bold text-center mt-10 mb-20'>Cart</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-10">
+            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-10 w-11/12 mx-auto">
                 {
                     loadedItems.map(item => (
                         <div key={item._id} className="card card-compact w-96 bg-base-100 shadow-xl">
